gmail_connector/requirements.txt
google-api-python-client>=2.120.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
psycopg2-binary>=2.9
redis>=5.0
rq>=1.16
requests>=2.32
python-dotenv>=1.0
